Extract package 'rest'
